requirements.txt
shortuuid
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.7
Markdown==3.0.1
pycparser==2.19
PyJWT==1.7.0
PyMySQL==0.9.2
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
gunicorn==19.7.1
gevent==1.3.3